refactor: 디렉토리 구조 변경
1. 컴포넌트는 components 폴더에 분리
2. RestaurantValidator를 도메인으로 이동
